+++ huskarl-resource-server/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+ - Breaking: Make AccessTokenValidator dyn-compatible (with boxed futures).
+ - Add MultiIssuerValidator which can route validator by iss value.
+ - Add `non_exhaustive` to various error types.
+ - Breaking: Add Debug bound to ToRfc6750Error.
+ 
 ## [0.8.0] - 2026-06-15
 
  - Breaking: ported to the dyn-first huskarl-core.